delete malformed annotations in commitment_bank
2962_clipped_2 and 2962_clipped_3:
    these have the same cb_target and cb_vals but the audio is DIFFERENT.
3561_clipped and 3561_clipped_1:
    these have same cb_target but different cb_vals and the audio is identical.

When we went back to De Marneffe's data we found the correct cb_val
and audio corresponded to 2962_clipped_2 and so we deleted 2962_clipped_3 which
did not have an entry in their spreadsheet.

When we went back to look at 3561_clipped and 3561_clipped_1, what we
found was more confusing. The exact same example just has 2 different
aggregated (averaged over multiple annotators) cb_val ratings and appears
twice in their spreadsheet. It is unclear what to do in this case.
Presumably one of them is correct but since we cannot disambiguate we
chose to delete both entries.

So in summary the actions taken were:

1. DELETE 2962_clipped_3 - spurious inclusion.
2. DELETE 3561_clipped - ambiguous data error.
3. DELETE 3561_clipped_1 - ambiguous data error.